give guide examples its own js file
